week 18 - knowledge graph/requirements.txt
numpy
scipy
scikit-learn
matplotlib
graphiti-core
python-dotenv
# langchain
langchain
langgraph
langsmith
langchain_core
langchain_community
langchain_text_splitters
langchain_chroma
langchain_tavily
langchain-ollama
langchainhub
# langchain dependencies
pydantic
tiktoken
chromadb
faiss-cpu
unstructured[md]
python-magic